+++ servicecatalog_puppet/asset_helpers_unit_test.py
@@ -29,14 +29,13 @@ def test_resolve_from_site_packages(sut):
 def test_read_from_site_packages(sut, mocker):
     # setup
     mocked_open = mocker.patch.object(sut, 'open')
-    expected_result = 'hello world'
     expected_param = os.path.sep.join([
         os.path.dirname(os.path.abspath(__file__)),
         'foo'
     ])
 
     # exercise
-    actual_result = sut.read_from_site_packages('foo')
+    sut.read_from_site_packages('foo')
 
     # verify
     assert mocked_open.call_count == 1
