+++ servicecatalog_puppet/sdk.py
@@ -1,7 +1,7 @@
 # Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-from . import cli_commands
+import core
 
 
 def run(what="puppet", wait_for_completion=False):
@@ -11,7 +11,7 @@ def run(what="puppet", wait_for_completion=False):
     :param what: what should be run.  The only parameter that will work is ``puppet``
     :param wait_for_completion: Whether the command should wait for the completion of the pipeline before it returns
     """
-    cli_commands.run(what, wait_for_completion)
+    core.run(what, wait_for_completion)
 
 
 def add_to_accounts(account_or_ou):
@@ -20,7 +20,7 @@ def add_to_accounts(account_or_ou):
 
     :param account_or_ou: A dict describing the the account or the ou to be added
     """
-    cli_commands.add_to_accounts(account_or_ou)
+    core.add_to_accounts(account_or_ou)
 
 
 def remove_from_accounts(account_id_or_ou_id_or_ou_path):
@@ -30,7 +30,7 @@ def remove_from_accounts(account_id_or_ou_id_or_ou_path):
     :param account_id_or_ou_id_or_ou_path: the value can be an account_id, ou_id or an ou_path.  It should be present \
     in the accounts list within the manifest file or an error is generated
     """
-    cli_commands.remove_from_accounts(account_id_or_ou_id_or_ou_path)
+    core.remove_from_accounts(account_id_or_ou_id_or_ou_path)
 
 
 def add_to_launches(launch_name, launch):
@@ -40,7 +40,7 @@ def add_to_launches(launch_name, launch):
     :param launch_name: The launch name to use when adding the launch to the manifest launches
     :param launch: The dict to add to the launches
     """
-    cli_commands.add_to_launches(launch_name, launch)
+    core.add_to_launches(launch_name, launch)
 
 
 def remove_from_launches(launch_name):
@@ -49,7 +49,7 @@ def remove_from_launches(launch_name):
 
     :param launch_name: The name of the launch to be removed from the launches section of the manifest file
     """
-    cli_commands.remove_from_launches(launch_name)
+    core.remove_from_launches(launch_name)
 
 
 def upload_config(config):
@@ -63,7 +63,7 @@ def upload_config(config):
 
     :param config: The dict containing the configuration used for puppet
     """
-    cli_commands.upload_config(config)
+    core.upload_config(config)
 
 
 def bootstrap(with_manual_approvals):
@@ -73,7 +73,7 @@ def bootstrap(with_manual_approvals):
 
     :param with_manual_approvals: Boolean to specify whether there should be manual approvals before provisioning occurs
     """
-    cli_commands.bootstrap(with_manual_approvals)
+    core.bootstrap(with_manual_approvals)
 
 
 def bootstrap_spoke(puppet_account_id):
@@ -83,7 +83,7 @@ def bootstrap_spoke(puppet_account_id):
 
     :param puppet_account_id: this is the account id where you have installed aws-service-catalog-puppet
     """
-    cli_commands.bootstrap_spoke(puppet_account_id)
+    core.bootstrap_spoke(puppet_account_id)
 
 
 def bootstrap_spoke_as(puppet_account_id, iam_role_arns):
@@ -95,7 +95,7 @@ def bootstrap_spoke_as(puppet_account_id, iam_role_arns):
     :param iam_role_arns: this is a list of ARNs the function will assume (in order) before bootstrapping.  The final \
     ARN in the list should be the ARN of the spoke you want to bootstrap.
     """
-    cli_commands.bootstrap_spoke_as(puppet_account_id, iam_role_arns)
+    core.bootstrap_spoke_as(puppet_account_id, iam_role_arns)
 
 
 def bootstrap_spokes_in_ou(ou_path_or_id, role_name, iam_role_arns):
@@ -107,4 +107,4 @@ def bootstrap_spokes_in_ou(ou_path_or_id, role_name, iam_role_arns):
     :param iam_role_arns: this is a list of ARNs the function will assume (in order) before bootstrapping.  The final \
     ARN in the list should be the ARN of account that can assume the role_name in the accounts to bootstrap.
     """
-    cli_commands.bootstrap_spokes_in_ou(ou_path_or_id, role_name, iam_role_arns)
+    core.bootstrap_spokes_in_ou(ou_path_or_id, role_name, iam_role_arns)
