+++ servicecatalog_puppet/workflow/provisioning_unit_test.py
@@ -5,8 +5,8 @@ from pytest import fixture
 
 @fixture
 def module():
-    from . import luigi_tasks_and_targets
-    return luigi_tasks_and_targets
+    from . import provisioning
+    return provisioning
 
 
 @fixture
@@ -114,7 +114,7 @@ class TestProvisionProductTask():
             self, module, minimal_params, mocker
     ):
         # setup
-        mocker.patch.object(module, 'cli_command_helpers')
+        mocker.patch.object(module, 'config')
         expected_ssm_params = {
             'Foo': {
             }
@@ -144,7 +144,7 @@ class TestProvisionProductTask():
             portfolio, product, version, account_id, region
     ):
         # setup
-        mocker.patch.object(module, 'cli_command_helpers')
+        mocker.patch.object(module, 'config')
         sut = module.ProvisionProductTask(**minimal_params)
 
         # exercise
